提交一下代码，修改一下警告。和cg的AddReference

+++ WLEngine/AllocationHeader.h
@@ -37,7 +37,7 @@ namespace WL
 #endif
 	{
 	public:
-		UINT16 getAllocatorIdentifier() const
+		UINT32 getAllocatorIdentifier() const
 		{
 			return m_AllocatorIdentifier; 
 		}
